requirements.txt
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
apprise==1.9.2
asgi-lifespan==2.1.0
asgiref==3.8.1
asttokens==3.0.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.3.0
black==24.10.0
blis==1.2.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.1
colorama==0.4.6
confection==0.1.5
coolname==2.2.0
croniter==6.0.0
cryptography==44.0.0
cymem==2.0.11
dateparser==1.2.0
decorator==5.1.1
Deprecated==1.2.15
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.16
email_validator==2.2.0
en_core_web_lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
fsspec==2024.12.0
graphviz==0.20.3
greenlet==3.1.1
griffe==1.5.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
htmltools==0.6.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
ipython==8.31.0
jedi==0.19.2
jellyfish==1.1.3
Jinja2==3.1.5
jinja2-humanize-extension==0.4.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langcodes==3.5.0
language_data==1.3.0
linkify-it-py==2.0.3
Mako==1.3.8
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.22.0
nltk==3.9.1
numpy==2.2.2
oauthlib==3.2.2
opentelemetry-api==1.29.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
piccolo==1.22.0
platformdirs==4.3.6
prefect==3.1.13
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.5
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-slugify==8.0.4
python-socks==2.6.1
pytz==2024.2
PyYAML==6.0.2
questionary==2.1.0
readchar==4.2.1
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
shellingham==1.5.4
shiny==1.2.1
six==1.17.0
skillNer==1.0.3
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
srsly==2.5.1
stack-data==0.6.3
starlette==0.41.3
targ==0.4.0
text-unidecode==1.3
thinc==8.3.4
time-machine==2.16.0
toml==0.10.2
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uc-micro-py==1.0.3
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
websockets==13.1
wrapt==1.17.2
zipp==3.21.0
